Correction 1

+++ 0x01-python_async_function/1-concurrent_coroutines.py
@@ -8,6 +8,6 @@ from typing import List
 wait_random = __import__('0-basic_async_syntax').wait_random
 
 
-async def wait_n(n: int, max_delay: int) -> List[int]:
+async def wait_n(n: int, max_delay: int) -> List[float]:
     """ Waiting async routine """
     return await asyncio.gather(*[wait_random(max_delay) for _ in range(n)])
